requirements.txt
Django==1.6
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-annoying==0.7.7
django-dbarray==0.0.1
django-extensions==1.2.5
django-model-utils==1.5.0
django-toolbelt==0.0.1
djorm-ext-core==0.4.3
djorm-ext-expressions==0.6
djorm-ext-pgarray==0.8
gunicorn==18.0
psycopg2==2.5.1
six==1.4.1
static==0.4
wsgiref==0.1.2
pyparsing==1.5.7
pydot==1.0.28

